Fix warnings about comments at the top of interfaces.
The `@typescript-eslint/eslint-plugin` plugin recently gained a fix for
the spurious lines-around-comments warnings at the tops of interfaces
and other typescript-specific blocks. This change updates our plugin to
depend on that, and adds the new configuration to enable them.

Testing Done:
Set up with a build of this and saw that I was no longer getting
warnings about comments that were the first line of an interface.

Reviewed at https://reviews.reviewboard.org/r/13008/