Update our ESLint configuration for JSX, Storybook, and template literals.
This updates our ESLint configuration to enable JSX parsing for any
`*.jsx` or `*.tsx` files. While we don't use React in our code, JSX may
be a useful option in the future, and is used now for Storybook, a UI
component library viewer.

It also adds some rules specific to Storybook "Story" files
(`*.stories.{js,jsx,ts,tsx}`). Specifically, it disables sorted keys in
objects, as that makes it difficult to maintain standard Story files,
where the order of keys can be important.

And last, it disables indentation rules in template literals. These
rules kicked in when trying to embed something that spanned multiple
lines, which is more common in JSX-like grammars (which we'll be using
in Ink).

Testing Done:
Made use of the new rules in the development of Ink.

Reviewed at https://reviews.reviewboard.org/r/13570/