Bump and add TypeScript-related dependencies.
This bumps our `typescript-eslint` dependency to the latest version,
7.1.0. This contains support for modern TypeScript, as well as a
plethora of fixes and rules that were introduced since our now aging
5.59.2 release.

It also adds an explicit development dependency on TypeScript, ensuring
that we're consistent in what we depend on within the codebase for
linting. This avoids assumptions that can be made by the Babel support
and warnings that may be emitted when running ESLint within this
codebase.

Testing Done:
Successfully linted and built the codebase.

Reviewed at https://reviews.reviewboard.org/r/13574/