Allow us to define empty interfaces for extends.
For some of our attributes definitions with inherited models/views, it
can be handy to define an empty interface that extends the base class
one. This change enables that for our linting rules.

I've also alphabetized the rules in this definition in case it expands
further.

Testing Done:
Set up a test case where we defined an interface that did nothing but
extend a base interface, and saw that I no longer got errors about it
from eslint.

Reviewed at https://reviews.reviewboard.org/r/14680/